util/requirements.txt
pyexiftool
argparse